requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crewai==0.22.5
curl_cffi==0.6.2
dataclasses-json==0.6.4
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.15
duckduckgo_search==5.1.0
fastapi==0.110.0
frozenlist==1.4.1
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
instructor==0.5.2
jaraco.context==4.3.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.31
lxml==5.1.0
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
setuptools==69.2.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.18.1
